StylingAndFeatures

+++ client/src/components/message/MessageDisplay.js
@@ -1,42 +1,55 @@
 import React from "react";
-import { Card, CardsContainer, Button, Li, P } from "../../styles";
+import { Card, CardsContainer, CardBody, Li, P } from "../../styles";
 import { deleteMessage } from "../../actions/message";
 import { connect } from "react-redux";
+import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
+import { faTrash } from "@fortawesome/free-solid-svg-icons";
+import { getReceiverOrSenderName } from "./helper";
 
-const MessagesDisplay = ({ messages, type, deleteMessage }) => {
+const MessagesDisplay = ({
+  messages,
+  type,
+  deleteMessage,
+  users: { usersList },
+}) => {
   const handleDelete = async (messageId) => {
-    await deleteMessage(messageId);
+    if (window.confirm("Are you sure you want to delete this message?"))
+      await deleteMessage(messageId);
   };
+
   return (
     <CardsContainer>
       {messages.map((message, index) => {
+        const name = getReceiverOrSenderName(usersList, type, message);
         return (
           <Card
             key={index}
-            title={`User Id : ${
-              type === "received"
-                ? `${message.senderId}`
-                : `${message.receiverId}`
-            }  `}
-            headStyle={{ borderBottom: "1px solid #c7c7c7" }}
+            title={name}
+            headStyle={{
+              borderBottom: "1px solid #c7c7c7",
+            }}
           >
             <>
               {" "}
-              <div>
+              <CardBody>
                 <Li>
                   {<P>{message.subject}</P>}
                   <ul>{message.message}</ul>
-                  <ul>{new Date(message.createdAt).toLocaleString()}</ul>
+                  <ul style={{ fontSize: "medium" }}>
+                    {new Date(message.createdAt).toLocaleString()}
+                  </ul>
                 </Li>
-              </div>
-              <Button
-                onClick={() => {
-                  handleDelete(message._id);
-                }}
-                isDangerButton
-              >
-                Delete
-              </Button>
+                <FontAwesomeIcon
+                  style={{ alignSelf: "center" }}
+                  icon={faTrash}
+                  color={"black"}
+                  size={"2x"}
+                  cursor="pointer"
+                  onClick={() => {
+                    handleDelete(message._id);
+                  }}
+                />{" "}
+              </CardBody>
             </>
           </Card>
         );
@@ -45,4 +58,8 @@ const MessagesDisplay = ({ messages, type, deleteMessage }) => {
   );
 };
 
-export default connect(null, { deleteMessage })(MessagesDisplay);
+const mapStateToProps = (state) => ({
+  users: state.users,
+});
+
+export default connect(mapStateToProps, { deleteMessage })(MessagesDisplay);
